socks5 measurement proxy for OONI
The code added by this initial commit implements a simple
socks5 server that could be used to measure what happens in
the network by letting OONI use such proxy.

It is important to remark that this code is not meant to be
a fancy implementation. Quite the contrary, one objective is
for it to be simple enough that we can evolve it easily.

For this reason, I've avoided asynchronous I/O. We may use
a lot of threads (but OONI will need few). Anyway, with
the higher limits specified in the testing program, I have
been running the proxy with Firefox input for hours now
(of course not being idle) and the amount of CPU and memory
consumed seems to be reasonable. (Below than 20 MiB with
200 threads that often are blocked on I/O.)

Some notable features of this implementation are:

1. it tells you about any network error that occurs in
   system calls that we care about, normalizing the errors
   so that we only focus on a narrow set

2. if support for getdns is enabled, domain name resolution
   may optionally use getdns rather than getaddrinfo, thus
   exposing in principle a wealth of information

3. there is code sniffing the first bytes send to us by the
   server that extracts the TLS certificate chain sent by
   the server to the client and passes it to a callback

As a fun fact, I did not plan to write this code. I was thinking
about going to the supermarket. However, I decided to do some
tinkering with libuv to write an async socks5 server and noticed
that it was hard. Using some code from libndt (with some bits
of simplifications), I rewrote the code as sync. Now more than
24 hours have passed [*] and it's probably time for me to really
exit and go to the supermarket. So, I think that it's probably
a good moment to publish this on GitHub.

[*] No need to worry, I have a seven hours sleep in the middle.