Rewrite TLS processing code to cope with TLSv1.3
After a conversation with @darkk, I realised that the previous code
wasn't working with TLSv1.3, because the certificate is encrypted
with such protocol. Therefore, it seems that the correct way to get
the certificate is most likely asking cURL for it [1].

Because of that, in this commit I have rewritten the TLS tracking
code to follow the TLS stream and I have made it disabled when you
request tracing (i.e. you need to request tracing and after that
also enable tracing of TLS). I'm not removing the code right away
because it MAY still be usful with lower versions of TLS.

Even if this commit adds more lines than it deleted, most of them
are either declarations of comments. The new algorithm is equal
if not more simple than the previous one.